order-service/requirements.txt
Flask==3.0.3
pymongo==4.7.2
python-dotenv==1.0.1
redis==5.0.4